Fix: prevent XSS by sanitizing city input in server and sanitize.js (#296)
